Move dataset catalog provider to its own file
